FEATURE: Create IE Support Plugin
This core plugin, which could be split off in the future, allows us to load IE specific code on demand.
