+++ .gitignore
@@ -53,6 +53,7 @@ bootsnap-compile-cache/
 !/plugins/discourse-narrative-bot
 !/plugins/discourse-presence
 !/plugins/discourse-local-dates
+!/plugins/discourse-internet-explorer
 /plugins/*/auto_generated/
 
 /spec/fixtures/plugins/my_plugin/auto_generated
