+++ plugins/discourse-internet-explorer/assets/stylesheets/ie.scss
@@ -1,0 +1,38 @@
+@media screen and (max-width: 767px) {
+  table.staff-logs tr {
+    display: -ms-grid;
+    -ms-grid-columns: 1fr 1fr 1fr 0.5fr;
+    -ms-grid-rows: auto auto;
+    td {
+      display: -ms-grid;
+      &.staff-users {
+        -ms-grid-row: 1;
+        -ms-grid-column: 1;
+        -ms-grid-column-span: 2;
+      }
+      &.created-at {
+        -ms-grid-row: 1;
+        -ms-grid-column: 4;
+      }
+      &.action {
+        -ms-grid-row: 2;
+        -ms-grid-column: 1;
+      }
+      &.subject {
+        -ms-grid-row: 2;
+        -ms-grid-column: 2;
+        -ms-grid-column-span: 3;
+      }
+      &.details {
+        -ms-grid-row: 3;
+        -ms-grid-column: 1;
+        -ms-grid-column-span: 3;
+      }
+      &.context {
+        -ms-grid-row: 4;
+        -ms-grid-column: 1;
+        -ms-grid-column-span: 3;
+      }
+    }
+  }
+}
