+++ plugins/discourse-internet-explorer/config/locales/server.en.yml
@@ -1,0 +1,3 @@
+en:
+  site_settings:
+    discourse_internet_explorer_enabled: "Internet Explorer support"
