+++ plugins/discourse-internet-explorer/config/settings.yml
@@ -1,0 +1,3 @@
+plugins:
+  discourse_internet_explorer_enabled:
+    default: true
